pyproject.toml
[project]
name = "kashflo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.120.1",
    "passlib>=1.7.4",
    "psycopg2>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
